pyproject.toml
[project]
name = "autoedge-agentic-benchmark"
version = "0.1.0"
description = "ITBench-Lite agentic benchmark across model families on Ollama"
requires-python = ">=3.11"
dependencies = [
  "beautifulsoup4>=4.12.0",
  "datasets>=4.0.0",
  "deep-translator>=1.11.4",
  "docling[asr]>=2.0.0",
  "huggingface-hub>=0.30.0",
  "matplotlib>=3.9.0",
  "numpy>=2.0.0",
  "openpyxl>=3.1.0",
  "pandas>=2.2.0",
  "pillow>=10.0.0",
  "playwright>=1.54.0",
  "python-docx>=1.1.0",
  "python-pptx>=1.0.2",
  "pytest>=8.3.0",
  "pyyaml>=6.0.2",
  "requests>=2.31.0",
  "scipy>=1.15.0",
  "seaborn>=0.13.0",
  "tqdm>=4.66.0",
]

[tool.uv]
package = false
